# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Benefits Enrollment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /benefits:
    get:
      operationId: listBenefitPlans
      tags: [benefits]
      summary: List all benefit plans
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: carrierId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of benefit plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BenefitPlan"
    post:
      operationId: createBenefitPlan
      tags: [benefits]
      summary: Create a benefit plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BenefitPlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenefitPlan"
  /benefits/{id}:
    get:
      operationId: getBenefitPlan
      tags: [benefits]
      summary: Get benefit plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Benefit plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenefitPlan"
    patch:
      operationId: updateBenefitPlan
      tags: [benefits]
      summary: Update benefit plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BenefitPlanUpdate"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenefitPlan"
    delete:
      operationId: deleteBenefitPlan
      tags: [benefits]
      summary: Delete benefit plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Enrollments
...
  /carriers:
    get:
      operationId: listCarriers
      tags: [carriers]
      summary: List all carriers
      responses:
        "200":
          description: List of carriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Carrier"
    post:
      operationId: createCarrier
      tags: [carriers]
      summary: Create a carrier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierInput"
      responses:
        "201":
          description: Created carrier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
  /carriers/{id}:
    get:
      operationId: getCarrier
      tags: [carriers]
      summary: Get carrier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Carrier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
    patch:
      operationId: updateCarrier
      tags: [carriers]
      summary: Update carrier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarrierUpdate"
      responses:
        "200":
          description: Updated carrier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
    delete:
      operationId: deleteCarrier
      tags: [carriers]
      summary: Delete carrier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /carriers/{id}/test-connection:
    post:
      operationId: testCarrierConnection
      tags: [carriers]
      summary: Test carrier EDI/SFTP connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"

  # Documents
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: carrierId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create/upload a document record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /documents/{id}/transmit-email:
    post:
      operationId: transmitDocumentEmail
      tags: [documents]
      summary: Email document to carrier enrollment department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTransmitInput"
      responses:
        "200":
          description: Email transmit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransmitResult"

  # Integrations (Payroll systems)
...
components:
  schemas:
...
    BenefitPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        carrierId:
          type: integer
        carrierName:
          type: ["string", "null"]
        groupNumber:
          type: ["string", "null"]
        planYear:
          type: ["string", "null"]
        effectiveDate:
          type: ["string", "null"]
        renewalDate:
          type: ["string", "null"]
        employeeCost:
          type: ["number", "null"]
        employerCost:
          type: ["number", "null"]
        deductible:
          type: ["number", "null"]
        outOfPocketMax:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        enrolledCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - carrierId
        - isActive
        - enrolledCount
        - createdAt
    BenefitPlanInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        carrierId:
          type: integer
        groupNumber:
          type: string
        planYear:
          type: string
        effectiveDate:
          type: string
        renewalDate:
          type: string
        employeeCost:
          type: number
        employerCost:
          type: number
        deductible:
          type: number
        outOfPocketMax:
          type: number
        description:
          type: string
      required:
        - name
        - type
        - carrierId
    BenefitPlanUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        carrierId:
          type: integer
        groupNumber:
          type: string
        planYear:
          type: string
        effectiveDate:
          type: string
        renewalDate:
          type: string
        employeeCost:
          type: number
        employerCost:
          type: number
        deductible:
          type: number
        outOfPocketMax:
          type: number
        description:
          type: string
        isActive:
          type: boolean
...
    TransmitResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        referenceId:
          type: ["string", "null"]
      required:
        - success
        - message
    Carrier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        enrollmentEmail:
          type: ["string", "null"]
        ediEnabled:
          type: boolean
        ediSenderId:
          type: ["string", "null"]
        ediReceiverId:
          type: ["string", "null"]
        sftpEnabled:
          type: boolean
        sftpHost:
          type: ["string", "null"]
        sftpPort:
          type: ["integer", "null"]
        sftpUsername:
          type: ["string", "null"]
        sftpPath:
          type: ["string", "null"]
        apiEnabled:
          type: boolean
        apiEndpoint:
          type: ["string", "null"]
        connectionStatus:
          type: string
        lastTestedAt:
          type: ["string", "null"]
        activePlans:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - ediEnabled
        - sftpEnabled
        - apiEnabled
        - connectionStatus
        - activePlans
        - createdAt
    CarrierInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        enrollmentEmail:
          type: string
        ediEnabled:
          type: boolean
        ediSenderId:
          type: string
        ediReceiverId:
          type: string
        sftpEnabled:
          type: boolean
        sftpHost:
          type: string
        sftpPort:
          type: integer
        sftpUsername:
          type: string
        sftpPassword:
          type: string
        sftpPath:
          type: string
        apiEnabled:
          type: boolean
        apiEndpoint:
          type: string
        apiKey:
          type: string
      required:
        - name
        - type
    CarrierUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        enrollmentEmail:
          type: string
        ediEnabled:
          type: boolean
        ediSenderId:
          type: string
        ediReceiverId:
          type: string
        sftpEnabled:
          type: boolean
        sftpHost:
          type: string
        sftpPort:
          type: integer
        sftpUsername:
          type: string
        sftpPassword:
          type: string
        sftpPath:
          type: string
        apiEnabled:
          type: boolean
        apiEndpoint:
          type: string
        apiKey:
          type: string
    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: ["integer", "null"]
      required:
        - success
        - message
    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        carrierId:
          type: ["integer", "null"]
        carrierName:
          type: ["string", "null"]
        fileName:
          type: string
        fileSize:
          type: ["integer", "null"]
        isPdfFillable:
          type: boolean
        transmissionStatus:
          type: string
        lastTransmittedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - fileName
        - isPdfFillable
        - transmissionStatus
        - createdAt
    DocumentInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        carrierId:
          type: integer
        fileName:
          type: string
        fileSize:
          type: integer
        isPdfFillable:
          type: boolean
        notes:
          type: string
      required:
        - name
        - type
        - fileName
    EmailTransmitInput:
      type: object
      properties:
        recipientEmail:
          type: string
        subject:
          type: string
        message:
          type: string
      required:
        - recipientEmail
...