# lib/api-zod/src/generated/types/carrier.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Benefits Enrollment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Carrier {
  id: number;
  name: string;
  type: string;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  enrollmentEmail?: string | null;
  ediEnabled: boolean;
  /** @nullable */
  ediSenderId?: string | null;
  /** @nullable */
  ediReceiverId?: string | null;
  sftpEnabled: boolean;
  /** @nullable */
  sftpHost?: string | null;
  /** @nullable */
  sftpPort?: number | null;
  /** @nullable */
  sftpUsername?: string | null;
  /** @nullable */
  sftpPath?: string | null;
  apiEnabled: boolean;
  /** @nullable */
  apiEndpoint?: string | null;
  connectionStatus: string;
  /** @nullable */
  lastTestedAt?: string | null;
  activePlans: number;
  createdAt: string;
}