# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Benefits Enrollment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  name: string;
  type: string;
  /** @nullable */
  carrierId?: number | null;
  /** @nullable */
  carrierName?: string | null;
  fileName: string;
  /** @nullable */
  fileSize?: number | null;
  isPdfFillable: boolean;
  transmissionStatus: string;
  /** @nullable */
  lastTransmittedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}