# lib/api-zod/src/generated/types/transmitResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Benefits Enrollment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TransmitResult {
  success: boolean;
  message: string;
  /** @nullable */
  referenceId?: string | null;
}